- adding User

+++ app/src/main/java/com/scp/adminforquiz/model/api/NwUser.java
@@ -4,10 +4,9 @@ import com.google.gson.annotations.SerializedName;
 
 public class NwUser {
     @SerializedName("id")
-    public Integer id;
+    public Long id;
     @SerializedName("fullName")
     public String fullName;
     @SerializedName("avatar")
     public String avatar;
-
 }
